// src/lib/types/playlistPanelVideoRenderer.ts
export interface IPlaylistPanelVideoRenderer {
	title: ShortBylineText
	longBylineText: LongBylineText
	thumbnail: IPlaylistPanelVideoRendererThumbnail
	lengthText: LengthText
	selected: boolean
	navigationEndpoint: IPlaylistPanelVideoRendererNavigationEndpoint
	videoId: string
	shortBylineText: ShortBylineText
	trackingParams: string
	menu: Menu
	playlistSetVideoId: string
}

interface LengthText {
	runs: LengthTextRun[]
	accessibility: Accessibility
}

interface Accessibility {
	accessibilityData: AccessibilityData
}

interface AccessibilityData {
	label: string
}

interface LengthTextRun {
	text: string
}

interface LongBylineText {
	runs: LongBylineTextRun[]
}

interface LongBylineTextRun {
	text: string
	navigationEndpoint?: RunNavigationEndpoint
}
// ...
interface Menu {
	menuRenderer: MenuRenderer
}

interface MenuRenderer {
	items: ItemElement[]
	trackingParams: string
	accessibility: Accessibility
}
// ...
interface ShortBylineText {
	runs: LengthTextRun[]
}
// ...
interface IPlaylistPanelVideoRendererNavigationEndpoint {
	clickTrackingParams: string
	watchEndpoint: FluffyWatchEndpoint
}

interface FluffyWatchEndpoint {
	videoId: string
	playlistId: string
	index: number
	params: string
	playerParams: string
	playlistSetVideoId: string
	loggingContext: LoggingContext
	watchEndpointMusicSupportedConfigs: FluffyWatchEndpointMusicSupportedConfigs
}

interface FluffyWatchEndpointMusicSupportedConfigs {
	watchEndpointMusicConfig: FluffyWatchEndpointMusicConfig
}

interface FluffyWatchEndpointMusicConfig {
	hasPersistentPlaylistPanel: boolean
	musicVideoType: string
}

interface IPlaylistPanelVideoRendererThumbnail {
	thumbnails: ThumbnailElement[]
}

interface ThumbnailElement {
	url: string
	width: number
	height: number
}